+++ libs/ballot-interpreter-vx/src/interpreter.dual_language_ballots.test.ts
@@ -1,0 +1,45 @@
+import * as hamilton from '../test/fixtures/election-5c6e578acf-state-of-hamilton-2020';
+import { Interpreter } from './interpreter';
+
+jest.setTimeout(10000);
+
+test('dual language ballot', async () => {
+  const { election } = hamilton;
+  const interpreter = new Interpreter({ election });
+
+  await interpreter.addTemplate(
+    await interpreter.interpretTemplate(
+      await hamilton.blankPage1.imageData(),
+      await hamilton.blankPage1.metadata()
+    )
+  );
+
+  const { ballot } = await interpreter.interpretBallot(
+    await hamilton.filledInPage1.imageData()
+  );
+  expect(ballot.votes).toMatchInlineSnapshot(`
+    Object {
+      "president": Array [
+        Object {
+          "id": "barchi-hallaren",
+          "name": "Joseph Barchi and Joseph Hallaren",
+          "partyId": "0",
+        },
+      ],
+      "representative-district-6": Array [
+        Object {
+          "id": "schott",
+          "name": "Brad Schott",
+          "partyId": "2",
+        },
+      ],
+      "senator": Array [
+        Object {
+          "id": "brown",
+          "name": "David Brown",
+          "partyId": "6",
+        },
+      ],
+    }
+  `);
+});
