+++ libs/ballot-interpreter-vx/src/utils/template_map.test.ts
@@ -1,12 +1,13 @@
-import { BallotType, BallotPageLayout } from '@votingworks/types';
+import { BallotType, BallotPageLayoutWithImage } from '@votingworks/types';
 import { createImageData } from './canvas';
 import { templateMap, TemplateMapKey } from './template_map';
 
 test('no locale info', () => {
   const map = templateMap();
-  const layout: BallotPageLayout = {
-    ballotImage: {
-      imageData: createImageData(1, 1),
+  const layout: BallotPageLayoutWithImage = {
+    imageData: createImageData(1, 1),
+    ballotPageLayout: {
+      pageSize: { width: 1, height: 1 },
       metadata: {
         ballotStyleId: '1',
         ballotType: BallotType.Standard,
@@ -16,8 +17,8 @@ test('no locale info', () => {
         pageNumber: 1,
         precinctId: '2',
       },
+      contests: [],
     },
-    contests: [],
   };
   const key: TemplateMapKey = [undefined, '1', '2', 1];
   map.set(key, layout);
@@ -27,9 +28,10 @@ test('no locale info', () => {
 
 test('with primary-only locale info', () => {
   const map = templateMap();
-  const layout: BallotPageLayout = {
-    ballotImage: {
-      imageData: createImageData(1, 1),
+  const layout: BallotPageLayoutWithImage = {
+    imageData: createImageData(1, 1),
+    ballotPageLayout: {
+      pageSize: { width: 1, height: 1 },
       metadata: {
         ballotStyleId: '1',
         ballotType: BallotType.Standard,
@@ -39,8 +41,8 @@ test('with primary-only locale info', () => {
         pageNumber: 1,
         precinctId: '2',
       },
+      contests: [],
     },
-    contests: [],
   };
   const key: TemplateMapKey = [{ primary: 'en-US' }, '1', '2', 1];
   map.set(key, layout);
@@ -50,9 +52,10 @@ test('with primary-only locale info', () => {
 
 test('with primary+secondary locale info', () => {
   const map = templateMap();
-  const layout: BallotPageLayout = {
-    ballotImage: {
-      imageData: createImageData(1, 1),
+  const layout: BallotPageLayoutWithImage = {
+    imageData: createImageData(1, 1),
+    ballotPageLayout: {
+      pageSize: { width: 1, height: 1 },
       metadata: {
         ballotStyleId: '1',
         ballotType: BallotType.Standard,
@@ -62,8 +65,8 @@ test('with primary+secondary locale info', () => {
         pageNumber: 1,
         precinctId: '2',
       },
+      contests: [],
     },
-    contests: [],
   };
   const key: TemplateMapKey = [
     { primary: 'en-US', secondary: 'es-US' },
